fix(history): handle leading tool messages
