+++ computer_control/__init__.py
@@ -1,0 +1,22 @@
+from . import controller
+from . import client
+from .analysis import (
+    list_python_files,
+    read_file,
+    search_code,
+    summarize_codebase,
+)
+from .main import main, trim_history
+from .controller import save_image
+
+__all__ = [
+    "client",
+    "controller",
+    "list_python_files",
+    "read_file",
+    "search_code",
+    "summarize_codebase",
+    "main",
+    "trim_history",
+    "save_image",
+]
